Move security transport implementation into security module
The security transport currently still exists in x-pack core, even
though the rest of security was split out when x-pack was split into
modules. It appears this was due to some tests indirectly relying (and
not needing to) rely on adding the netty4plugin. This commit moves the
implementation, and the netty module copy that is made in x-pack, into
the security module, which is what actually registers the transport
implementation.
