+++ x-pack/plugin/core/build.gradle
@@ -39,11 +39,6 @@ dependencies {
 
   // security deps
   api 'com.unboundid:unboundid-ldapsdk:4.0.8'
-  api project(path: ':modules:transport-netty4')
-  api( project(path: ':plugins:transport-nio')) {
-    // TODO: core exclusion should not be necessary, since it is a transitive dep of all plugins
-    exclude group: "org.elasticsearch", module: "elasticsearch-core"
-  }
 
   testImplementation 'org.elasticsearch:securemock:1.2'
   testImplementation "org.elasticsearch:mocksocket:${versions.mocksocket}"
