+++ x-pack/plugin/security/build.gradle
@@ -15,8 +15,8 @@ archivesBaseName = 'x-pack-security'
 
 dependencies {
   compileOnly project(path: xpackModule('core'))
-  compileOnly project(path: ':modules:transport-netty4')
-  compileOnly project(path: ':plugins:transport-nio')
+  api project(path: ':modules:transport-netty4')
+  api project(path: ':plugins:transport-nio')
 
   testImplementation project(path: xpackModule('monitoring'))
   testImplementation project(path: xpackModule('spatial'))
